+++ packages/eyes-sdk-core/lib/config/Configuration.js
@@ -242,6 +242,12 @@ class Configuration {
     /** @type {Object} */
     this._visualGridOptions = undefined
 
+    /** @type {number[]|boolean} */
+    this._layoutBreakpoints = undefined
+
+    /** @type {boolean} */
+    this._disableBrowserFetching = undefined
+
     if (configuration) {
       this.mergeConfig(configuration)
     }
@@ -1311,8 +1317,17 @@ class Configuration {
     } else if (breakpoints.length === 0) {
       this._layoutBreakpoints = false
     } else {
-      this._layoutBreakpoints = Array.from(new Set(breakpoints)).sort((a, b) => (a > b ? 1 : -1))
+      this._layoutBreakpoints = Array.from(new Set(breakpoints)).sort((a, b) => (a < b ? 1 : -1))
     }
+    return this
+  }
+
+  getDisableBrowserFetching() {
+    return this._disableBrowserFetching
+  }
+
+  setDisableBrowserFetching(value) {
+    this._disableBrowserFetching = value
     return this
   }