+++ packages/eyes-sdk-core/lib/sdk/EyesVisualGrid.js
@@ -256,7 +256,11 @@ class EyesVisualGrid extends EyesCore {
           checkSettings.getLayoutBreakpoints(),
           this._configuration.getLayoutBreakpoints(),
         )
-        const snapshots = await this._takeDomSnapshots({breakpoints})
+        const disableBrowserFetching = TypeUtils.getOrDefault(
+          checkSettings.getDisableBrowserFetching(),
+          this._configuration.getDisableBrowserFetching(),
+        )
+        const snapshots = await this._takeDomSnapshots({breakpoints, disableBrowserFetching})
         const [{url}] = snapshots
         if (this.getCorsIframeHandle() === CorsIframeHandles.BLANK) {
           snapshots.forEach(CorsIframeHandler.blankCorsIframeSrcOfCdt)
@@ -300,10 +304,14 @@ class EyesVisualGrid extends EyesCore {
   /**
    * @param {CheckSettings<TElement, TSelector>} checkSettings
    */
-  async _takeDomSnapshots({breakpoints}) {
+  async _takeDomSnapshots({breakpoints, disableBrowserFetching}) {
     const browsers = this._configuration.getBrowsersInfo()
     if (!breakpoints) {
-      const snapshot = await takeDomSnapshot({driver: this._driver})
+      const snapshot = await takeDomSnapshot({
+        driver: this._driver,
+        browser: this._userAgent.getBrowser(),
+        disableBrowserFetching,
+      })
       return Array(browsers.length).fill(snapshot)
     }
     const widths = await Promise.all(
@@ -319,14 +327,22 @@ class EyesVisualGrid extends EyesCore {
     const snapshots = {}
     if (widths.includes(viewportSize.getWidth())) {
       this._logger.log(`taking dom snapshot for existing width ${viewportSize.getWidth()}`)
-      const snapshot = await takeDomSnapshot({driver: this._driver})
+      const snapshot = await takeDomSnapshot({
+        driver: this._driver,
+        browser: this._userAgent.getBrowser(),
+        disableBrowserFetching,
+      })
       snapshots[viewportSize.getWidth()] = snapshot
     }
     for (const width of widths) {
       if (snapshots[width]) continue
       this._logger.log(`taking dom snapshot for width ${width}`)
       await this._driver.setViewportSize({width, height: viewportSize.getHeight()})
-      const snapshot = await takeDomSnapshot({driver: this._driver})
+      const snapshot = await takeDomSnapshot({
+        driver: this._driver,
+        browser: this._userAgent.getBrowser(),
+        disableBrowserFetching,
+      })
       snapshots[width] = snapshot
     }
     await this._driver.setViewportSize(viewportSize)