+++ packages/sdk-shared/coverage-tests/custom/TestVisualGridRefererHeader.spec.js
@@ -6,6 +6,7 @@ const testServer = require('../../src/test-server')
 const {Target} = require(cwd)
 const spec = require(path.resolve(cwd, 'src/SpecDriver'))
 const {getEyes} = require('../../src/test-setup')
+const preprocessUrl = require('../util/url-preprocessor')
 
 describe('TestVisualGridRefererHeader', () => {
   let testServer1, testServer2
@@ -36,9 +37,9 @@ describe('TestVisualGridRefererHeader', () => {
   })
 
   it('send referer header', async () => {
-    const url = 'http://localhost:5555/cors.html'
+    const url = preprocessUrl('http://localhost:5555/cors.html')
     await spec.visit(driver, url)
-    const eyes = getEyes('VG')
+    const eyes = getEyes({isVisualGrid: true})
     await eyes.open(driver, 'VgFetch', ' VgFetch referer', {width: 800, height: 600})
     await eyes.check('referer', Target.window())
     await eyes.close()